Added an -a switch for 'auto tagging'.
This currently only creates two tags, but they are immensely helpful
in automating restores of snapshots to volumes.

If this feature is liked, I can extend it to do more api work for
adding instance_id and other metadata to tags, but did not want
to initially increase dependencies or complexity.

anthroprose@gmail.com